// src/@types/response/customer.d.ts
export interface ICustomer {
  customerId: string;
  name: string;
  phone: string;
  citizenId: string;
  dateOfBirth: string;
  email: string;
  email: string;
  gender: string;
  cardId: string;
  accumulatedPoints: number;
  accumulatedDate: string;
  valid: boolean;
  rankName: "Membership" | "Silver" | "Gold";
}